move decimal parsing to cast and improve impl
